requirements.txt
black==23.11.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0
flake8==6.1.0
pytest==7.4.3
discord-py-interactions==4.4.0
pre-commit
